# docs/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Document Center API",
    "summary": "Документ-центр ИП «Северный Контур»: проекты, документы, версии и аудит",
    "description": "\nВнутренний REST API ИП «Северный Контур» для управления документами по проектам.\n\n**Базовый путь публичного контракта:** `/api/v1`. Версия контракта (`v1`)\nне совпадает с версией приложения, указанной в заголовке документации.\n\n### Аутентификация\nПолучите токен операцией `POST /api/v1/auth/login` и передавайте его в заголовке\n`Authorization: Bearer <token>`. В Swagger UI нажмите **Authorize** и вставьте токен.\n\n### Роли\n- **admin** — полный доступ ко всем проектам, пользователям и журналу действий;\n- **manager** — создаёт проекты и выдаёт доступ участникам своих проектов;\n- **worker** — работает с документами в рамках выданного доступа `editor`;\n- **viewer** — только чтение в рамках выданного доступа.\n\n### Ошибки\nВсе ошибки возвращаются в формате RFC 9457 Problem Details\n(`application/problem+json`) с полями `type`, `title`, `status`, `detail`,\n`instance`; ошибки валидации дополнительно содержат массив `errors`.\n\n### Коллекции\nКоллекции поддерживают параметры `offset` (≥ 0) и `limit` (1…100, по умолчанию 20)\nи возвращают конверт `items`, `total`, `offset`, `limit`.\n",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Выполнить вход",
        "description": "Проверить учётные данные и выпустить токен доступа.\n\nОперация доступна без аутентификации. Полученный токен передаётся\nв заголовке ``Authorization: Bearer <token>``. Число попыток входа\nс одного адреса ограничено (по умолчанию 10 в минуту); при превышении\nвозвращается 429 с заголовком ``Retry-After``.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Токен доступа со схемой Bearer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "401": {
            "description": "Invalid email or password",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "User account is deactivated",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Получить текущего пользователя",
        "description": "Вернуть карточку пользователя, которому принадлежит токен.",
        "operationId": "get_current_user",
        "responses": {
          "200": {
            "description": "Карточка аутентифицированного пользователя",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "User account is deactivated",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/access": {
      "get": {
        "tags": [
          "Access"
        ],
        "summary": "Получить список доступов проекта",
        "description": "Вернуть участников проекта и выданные им уровни доступа.\n\nОперация доступна владельцу проекта и администратору.",
        "operationId": "list_project_access",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Идентификатор проекта",
              "title": "Project Id"
            },
            "description": "Идентификатор проекта"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Смещение от начала коллекции",
              "default": 0,
              "title": "Offset"
            },
            "description": "Смещение от начала коллекции"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Размер страницы, максимум 100",
              "default": 20,
              "title": "Limit"
            },
            "description": "Размер страницы, максимум 100"
          }
        ],
        "responses": {
          "200": {
            "description": "Страница записей о выданных доступах",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ProjectAccessReadWithUser_"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Access denied",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/access/{user_id}": {
      "put": {
        "tags": [
          "Access"
        ],
        "summary": "Выдать или изменить доступ участника",
        "description": "Назначить участнику уровень доступа к проекту.\n\nОперация идемпотентна: повторный вызов с тем же телом не изменяет\nсостояние. Код 201 возвращается, если доступ выдан впервые, код 200 —\nесли изменён существующий уровень доступа.",
        "operationId": "upsert_project_access",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Идентификатор проекта",
              "title": "Project Id"
            },
            "description": "Идентификатор проекта"
          },
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Идентификатор участника",
              "title": "User Id"
            },
            "description": "Идентификатор участника"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectAccessUpsert"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Запись о доступе участника к проекту",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectAccessReadWithUser"
                }
              }
            }
          },
          "201": {
            "description": "Доступ выдан впервые",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectAccessReadWithUser"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Access denied",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Access"
        ],
        "summary": "Отозвать доступ участника",
        "description": "Отозвать ранее выданный доступ участника к проекту.",
        "operationId": "revoke_project_access",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Идентификатор проекта",
              "title": "Project Id"
            },
            "description": "Идентификатор проекта"
          },
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Идентификатор участника",
              "title": "User Id"
            },
            "description": "Идентификатор участника"
          }
        ],
        "responses": {
          "204": {
            "description": "Доступ отозван"
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Access denied",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FieldError": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field",
            "description": "Путь к полю запроса",
            "examples": [
              "title"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Причина отклонения значения",
            "examples": [
              "String should have at least 3 characters"
            ]
          }
        },
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "title": "FieldError",
        "description": "Описание одной ошибки валидации поля."
      },
...
      "Page_ProjectAccessReadWithUser_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProjectAccessReadWithUser"
            },
            "type": "array",
            "title": "Items",
            "description": "Записи текущей страницы"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total",
            "description": "Общее количество записей, доступных запросу"
          },
          "offset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offset",
            "description": "Смещение текущей страницы от начала коллекции"
          },
          "limit": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "description": "Запрошенный размер страницы"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ],
        "title": "Page[ProjectAccessReadWithUser]",
        "description": "Страница коллекции с метаданными пагинации: items, total, offset, limit"
      },
...
      "Permission": {
        "type": "string",
        "enum": [
          "viewer",
          "editor"
        ],
        "title": "Permission",
        "description": "Уровень доступа к проекту: viewer (чтение) или editor (изменение)."
      },
      "ProblemDetails": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Стабильный URI типа ошибки"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Краткое название типа ошибки, не зависящее от экземпляра"
          },
          "status": {
            "type": "integer",
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Status",
            "description": "HTTP-код ответа"
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Описание конкретного случая возникновения ошибки"
          },
          "instance": {
            "type": "string",
            "title": "Instance",
            "description": "Путь запроса, при обработке которого возникла ошибка"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance"
        ],
        "title": "ProblemDetails",
        "description": "Тело ответа с ошибкой по RFC 9457.",
        "examples": [
          {
            "detail": "Project 42 does not exist or is unavailable",
            "instance": "/api/v1/projects/42",
            "status": 404,
            "title": "Project not found",
            "type": "https://docs.local/problems/project-not-found"
          }
        ]
      },
      "ProjectAccessReadWithUser": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Идентификатор записи о доступе"
          },
          "project_id": {
            "type": "integer",
            "title": "Project Id",
            "description": "Идентификатор проекта"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "Идентификатор участника"
          },
          "permission": {
            "$ref": "#/components/schemas/Permission",
            "description": "Уровень доступа"
          },
          "granted_by": {
            "type": "integer",
            "title": "Granted By",
            "description": "Идентификатор пользователя, выдавшего доступ"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Момент выдачи доступа",
            "examples": [
              "2026-09-09T08:15:30Z"
            ]
          },
          "user_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Email",
            "description": "Адрес участника"
          },
          "granter_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Granter Email",
            "description": "Адрес пользователя, выдавшего доступ"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "user_id",
          "permission",
          "granted_by",
          "created_at"
        ],
        "title": "ProjectAccessReadWithUser",
        "description": "Запись о доступе с адресами участника и выдавшего доступ."
      },
      "ProjectAccessUpsert": {
        "properties": {
          "permission": {
            "$ref": "#/components/schemas/Permission",
            "description": "Уровень доступа участника к проекту",
            "default": "viewer",
            "examples": [
              "editor"
            ]
          }
        },
        "type": "object",
        "title": "ProjectAccessUpsert",
        "description": "Тело запроса на выдачу или изменение доступа."
      },
...
      "Token": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token",
            "description": "Подписанный JWT; передаётся в заголовке Authorization",
            "examples": [
              "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoxfQ.signature"
            ]
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "description": "Схема токена",
            "default": "bearer",
            "examples": [
              "bearer"
            ]
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "Token",
        "description": "Токен доступа, выдаваемый при входе."
      },
...
      "UserLogin": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Адрес электронной почты",
            "examples": [
              "user@example.com"
            ]
          },
          "password": {
            "type": "string",
            "title": "Password",
            "description": "Пароль в открытом виде",
            "examples": [
              "Passw0rd123"
            ]
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "UserLogin",
        "description": "Учётные данные для входа."
      },
      "UserRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Идентификатор пользователя"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Адрес электронной почты"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole",
            "description": "Роль пользователя"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "Признак активности учётной записи"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Момент создания учётной записи",
            "examples": [
              "2026-09-09T08:15:30Z"
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "role",
          "is_active",
          "created_at"
        ],
        "title": "UserRead",
        "description": "Карточка пользователя.\n\nХэш пароля и иные внутренние атрибуты в контракт не входят."
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "admin",
          "manager",
          "worker",
          "viewer"
        ],
        "title": "UserRole",
        "description": "Роль пользователя: admin, manager, worker или viewer."
      },
...
      "ValidationProblemDetails": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Стабильный URI типа ошибки"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Краткое название типа ошибки, не зависящее от экземпляра"
          },
          "status": {
            "type": "integer",
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Status",
            "description": "HTTP-код ответа"
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Описание конкретного случая возникновения ошибки"
          },
          "instance": {
            "type": "string",
            "title": "Instance",
            "description": "Путь запроса, при обработке которого возникла ошибка"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array",
            "title": "Errors",
            "description": "Перечень полей, не прошедших валидацию"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance"
        ],
        "title": "ValidationProblemDetails",
        "description": "Тело ответа для ошибок валидации запроса (HTTP 422).\n\nРасширяет базовую структуру членом ``errors`` согласно STD-ERR-05.",
        "examples": [
          {
            "detail": "Request body does not match the expected schema",
            "errors": [
              {
                "field": "title",
                "message": "String should have at least 3 characters"
              }
            ],
            "instance": "/api/v1/projects",
            "status": 422,
            "title": "Request validation failed",
            "type": "https://docs.local/problems/validation-error"
          }
        ]
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}